refactor: restructure application into modular architecture
- Split models and schemas into auth and blog modules
- Move routes to dedicated route files
- Update imports and references
- Clean up main.py to use routers
- Rename auth.py to auth_util.py for clarity
