+++ app/auth/model.py
@@ -1,0 +1,22 @@
+from sqlalchemy import Column, Integer, DateTime, Boolean, func, String
+
+from app.database import Base
+
+
+class User(Base):
+
+    __tablename__ = "users"
+
+    # Primary Key
+    id = Column(Integer, primary_key=True, index=True)
+
+    # Fields
+    email = Column(String, unique=True, index=True)
+    hashed_password = Column(String)
+    first_name = Column(String, nullable=False)
+    last_name = Column(String, nullable=False)
+
+    # Additional Fields
+    created_at = Column(DateTime(timezone=True), server_default=func.now())
+    updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
+    is_active = Column(Boolean, default=True)
