+++ app/auth/schema.py
@@ -1,24 +1,4 @@
 from pydantic import BaseModel, EmailStr
-from datetime import datetime
-
-class BlogBase(BaseModel):
-    title: str
-    content: str
-
-class BlogCreate(BlogBase):
-    pass
-
-class BlogResponse(BlogBase):
-    id: int
-    created_at: datetime
-
-    model_config = {
-        "from_attributes": True
-    }
-
-
-class BlogUpdate(BlogBase):
-    pass
 
 
 # Signup payload
