--- app/models.py
@@ -1,41 +1,0 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, func, Boolean, ForeignKey
-from app.database import Base
-
-
-class Blog(Base):
-    __tablename__ = "blogs"
-
-
-    # Foreign Key
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
-
-    # Primary Key
-    id = Column(Integer, primary_key=True, index=True)
-
-    # Fields
-    title = Column(String(255), nullable=False)
-    content = Column(Text, nullable=False)
-
-    # Additional Fields
-    created_at = Column(DateTime(timezone=True), server_default=func.now())
-    updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
-    is_active = Column(Boolean, default=True)
-
-
-class User(Base):
-
-    __tablename__ = "users"
-
-    # Primary Key
-    id = Column(Integer, primary_key=True, index=True)
-
-    # Fields
-    email = Column(String, unique=True, index=True)
-    hashed_password = Column(String)
-    first_name = Column(String, nullable=False)
-    last_name = Column(String, nullable=False)
-
-    # Additional Fields
-    created_at = Column(DateTime(timezone=True), server_default=func.now())
-    updated_at = Column(DateTime(timezone=True), server_default=func.now(), onupdate=func.now())
-    is_active = Column(Boolean, default=True)
